requirements.txt
streamlit
langchain
langchain-community
huggingface-hub
transformers
sentence-transformers
faiss-cpu
pypdf
python-dotenv
